+++ ta4j-core/src/main/java/org/ta4j/core/BacktestExecutor.java
@@ -23,12 +23,14 @@
  */
 package org.ta4j.core;
 
-import java.util.ArrayList;
-import java.util.List;
-
+import org.ta4j.core.cost.CostModel;
+import org.ta4j.core.cost.ZeroCostModel;
 import org.ta4j.core.num.Num;
 import org.ta4j.core.reports.TradingStatement;
 import org.ta4j.core.reports.TradingStatementGenerator;
+
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * This class enables backtesting of multiple strategies and comparing them to
@@ -43,14 +45,23 @@ public class BacktestExecutor {
         this(series, new TradingStatementGenerator());
     }
 
+    public BacktestExecutor(BarSeries series, CostModel transactionCostModel, CostModel holdingCostModel) {
+        this(series, new TradingStatementGenerator(), transactionCostModel, holdingCostModel);
+    }
+
     public BacktestExecutor(BarSeries series, TradingStatementGenerator tradingStatementGenerator) {
-        this.seriesManager = new BarSeriesManager(series);
+        this(series, tradingStatementGenerator, new ZeroCostModel(), new ZeroCostModel());
+    }
+
+    public BacktestExecutor(BarSeries series, TradingStatementGenerator tradingStatementGenerator,
+            CostModel transactionCostModel, CostModel holdingCostModel) {
+        this.seriesManager = new BarSeriesManager(series, transactionCostModel, holdingCostModel);
         this.tradingStatementGenerator = tradingStatementGenerator;
     }
 
     /**
      * Executes given strategies and returns trading statements.
-     * 
+     *
      * @param strategies the strategies
      * @param amount     the amount used to open/close the position
      */
@@ -61,7 +72,7 @@ public class BacktestExecutor {
     /**
      * Executes given strategies with specified trade type to open the position and
      * return the trading statements.
-     * 
+     *
      * @param strategies the strategies
      * @param amount     the amount used to open/close the position
      * @param tradeType  the {@link Trade.TradeType} used to open the position