+++ ta4j-examples/src/main/java/ta4jexamples/backtesting/SimpleMovingAverageRangeBacktest.java
@@ -23,26 +23,26 @@
  */
 package ta4jexamples.backtesting;
 
-import java.util.ArrayList;
-import java.util.List;
-
-import org.ta4j.core.BacktestExecutor;
-import org.ta4j.core.BarSeries;
-import org.ta4j.core.BaseStrategy;
-import org.ta4j.core.Indicator;
-import org.ta4j.core.Rule;
-import org.ta4j.core.Strategy;
-import org.ta4j.core.Trade;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.ta4j.core.*;
 import org.ta4j.core.indicators.SMAIndicator;
 import org.ta4j.core.indicators.helpers.ClosePriceIndicator;
 import org.ta4j.core.num.DecimalNum;
 import org.ta4j.core.num.Num;
+import org.ta4j.core.reports.PerformanceReport;
+import org.ta4j.core.reports.PositionStatsReport;
+import org.ta4j.core.reports.TradingStatement;
 import org.ta4j.core.rules.OverIndicatorRule;
 import org.ta4j.core.rules.UnderIndicatorRule;
-
 import ta4jexamples.loaders.CsvBarsLoader;
 
+import java.util.ArrayList;
+import java.util.List;
+
 public class SimpleMovingAverageRangeBacktest {
+
+    private static final Logger LOG = LoggerFactory.getLogger(SimpleMovingAverageRangeBacktest.class);
 
     public static void main(String[] args) {
         BarSeries series = CsvBarsLoader.loadAppleIncSeries();
@@ -58,7 +58,10 @@ public class SimpleMovingAverageRangeBacktest {
             strategies.add(strategy);
         }
         BacktestExecutor backtestExecutor = new BacktestExecutor(series);
-        backtestExecutor.execute(strategies, DecimalNum.valueOf(50), Trade.TradeType.BUY);
+        List<TradingStatement> tradingStatements = backtestExecutor.execute(strategies, DecimalNum.valueOf(50),
+                Trade.TradeType.BUY);
+
+        LOG.info(printReport(tradingStatements));
     }
 
     private static Rule createEntryRule(BarSeries series, int barCount) {
@@ -71,5 +74,46 @@ public class SimpleMovingAverageRangeBacktest {
         Indicator<Num> closePrice = new ClosePriceIndicator(series);
         SMAIndicator sma = new SMAIndicator(closePrice, barCount);
         return new OverIndicatorRule(sma, closePrice);
+    }
+
+    private static String printReport(List<TradingStatement> tradingStatements) {
+        StringBuilder resultBuilder = new StringBuilder();
+        resultBuilder.append(System.lineSeparator());
+        for (TradingStatement statement : tradingStatements) {
+            resultBuilder.append(printStatementReport(statement));
+            resultBuilder.append(System.lineSeparator());
+        }
+
+        return resultBuilder.toString();
+    }
+
+    private static StringBuilder printStatementReport(TradingStatement statement) {
+        StringBuilder resultBuilder = new StringBuilder();
+        resultBuilder.append("######### ").append(statement.getStrategy().getName()).append(" #########")
+                .append(System.lineSeparator()).append(printPerformanceReport(statement.getPerformanceReport()))
+                .append(System.lineSeparator()).append(printPositionStats(statement.getPositionStatsReport()))
+                .append(System.lineSeparator()).append("###########################");
+        return resultBuilder;
+    }
+
+    private static StringBuilder printPerformanceReport(PerformanceReport report) {
+        StringBuilder resultBuilder = new StringBuilder();
+        resultBuilder.append("--------- performance report ---------").append(System.lineSeparator())
+                .append("total loss: ").append(report.getTotalLoss()).append(System.lineSeparator())
+                .append("total profit: ").append(report.getTotalProfit()).append(System.lineSeparator())
+                .append("total profit loss: " + report.getTotalProfitLoss()).append(System.lineSeparator())
+                .append("total profit loss percentage: ").append(report.getTotalProfitLossPercentage())
+                .append(System.lineSeparator()).append("---------------------------");
+        return resultBuilder;
+    }
+
+    private static StringBuilder printPositionStats(PositionStatsReport report) {
+        StringBuilder resultBuilder = new StringBuilder();
+        resultBuilder.append("--------- trade statistics report ---------").append(System.lineSeparator())
+                .append("loss trade count: ").append(report.getLossCount()).append(System.lineSeparator())
+                .append("profit trade count: ").append(report.getProfitCount()).append(System.lineSeparator())
+                .append("break even trade count: ").append(report.getBreakEvenCount()).append(System.lineSeparator())
+                .append("---------------------------");
+        return resultBuilder;
     }
 }