+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - **Num** removed Serializable
 
 ### Added
+- **Enhancement** Added possibility to use CostModels when backtesting with the BacktestExecutor
 
 
 ## 0.14 (released April 25, 2021)